Allow users to view other users' habit details and habit events

+++ src/app/src/main/java/com/cmput301f21t09/budgetprojectname/views/activities/ViewHabitActivity.java
@@ -14,6 +14,7 @@ import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.cmput301f21t09.budgetprojectname.MainActivity;
+import com.cmput301f21t09.budgetprojectname.services.AuthorizationService;
 import com.cmput301f21t09.budgetprojectname.views.fragments.HabitScheduleViewSelector;
 import com.cmput301f21t09.budgetprojectname.views.lists.HabitEventCustomList;
 import com.cmput301f21t09.budgetprojectname.R;
@@ -66,6 +67,16 @@ public class ViewHabitActivity extends AppCompatActivity {
     private TextView habitDate;
 
     /**
+     * Button for editing a habit
+     */
+    private Button editHabitBtn;
+
+    /**
+     * ImageButton for removing a habit
+     */
+    private ImageButton removeHabitBtn;
+
+    /**
      * List View for habit events
      */
     private ListView habitEventList;
@@ -79,9 +90,13 @@ public class ViewHabitActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_habit_screen);
 
-        // Get the habitID from the previous activity
+        // Retrieve the logged in user's userid
+        String currentUserId = AuthorizationService.getInstance().getCurrentUserId();
+
+        // Get the habitID and habitUID from the previous activity
         Intent intent = getIntent();
         String habitID = intent.getStringExtra("HABIT_ID");
+        String habitUID = intent.getStringExtra("HABIT_USERID");
 
         /* Habit Details */
 
@@ -90,6 +105,8 @@ public class ViewHabitActivity extends AppCompatActivity {
         habitDescription = (TextView) findViewById(R.id.habitDescription);
         habitDate = (TextView) findViewById(R.id.habitDate);
         habitTitleToolbar = findViewById(R.id.toolbar_title);
+        editHabitBtn = findViewById(R.id.editHabitButton);
+        removeHabitBtn = findViewById(R.id.view_habit_remove_button);
 
         // Set up the controller and set the views accordingly
         controller = HabitController.getEditHabitController(habitID);
@@ -131,31 +148,41 @@ public class ViewHabitActivity extends AppCompatActivity {
                 intent.putExtra("HABIT_EVENT_ID", selectedHEM.getID());
                 intent.putExtra("HABIT_ID", selectedHEM.getHabitID());
                 intent.putExtra("HABIT_TITLE", habitTitle.getText().toString());
+                intent.putExtra("HABIT_EVENT_USERID", habitUID);
                 startActivity(intent);
             }
 
         });
 
+        // Brings the user back to the previous screen
         ImageButton back = findViewById(R.id.view_habit_back_button);
         back.setOnClickListener(v -> {
             finish();
         });
 
-        // Brings the user to another activity to edit habit details
-        final Button editHabitBtn = findViewById(R.id.editHabitButton);
-        editHabitBtn.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(), DefineHabitActivity.class);
-                final String HABIT_ID = "HABIT_ID";
-                intent.putExtra(HABIT_ID, habitID);
-                startActivity(intent);
-            }
-        });
+        // Let the user edit the habit if the user owns this habit
+        if (currentUserId.equals(habitUID)) {
+            editHabitBtn.setOnClickListener(new View.OnClickListener() {
+                public void onClick(View v) {
+                    Intent intent = new Intent(getApplicationContext(), DefineHabitActivity.class);
+                    final String HABIT_ID = "HABIT_ID";
+                    intent.putExtra(HABIT_ID, habitID);
+                    startActivity(intent);
+                }
+            });
+        } else {
+            editHabitBtn.setVisibility(View.INVISIBLE);
+        }
 
-        ImageButton removeHabitBtn = findViewById(R.id.view_habit_remove_button);
-        removeHabitBtn.setOnClickListener(view -> {
-            controller.deleteModel();
-        });
+        // Let the user delete the habit if the user owns this habit
+        if (currentUserId.equals(habitUID)) {
+            removeHabitBtn.setOnClickListener(view -> {
+                controller.deleteModel();
+            });
+        } else {
+            removeHabitBtn.setVisibility(View.INVISIBLE);
+        }
+
     }
 
     /**
